# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/onboarding/status": {
            "get": {
                "description": "Get current onboarding progress status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "onboarding"
                ],
                "summary": "Get onboarding status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer jwt-token...",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.ResponseBody"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPError"
                        }
                    }
                }
            }
        },
        "/onboarding/step1": {
            "post": {
                "description": "Update company profile information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "onboarding"
                ],
                "summary": "Onboarding Step 1: Update company profile",
                "parameters": [
                    {
                        "description": "Company profile update request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/onboarding.step1Request"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Bearer jwt-token...",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.ResponseBody"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPError"
                        }
                    }
                }
            }
        },
        "/onboarding/step2": {
            "post": {
                "description": "Create or update multiple users for the company in a single request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "onboarding"
                ],
                "summary": "Onboarding Step 2: Create/update users",
                "parameters": [
                    {
                        "description": "Create/update users request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/onboarding.step2Request"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Bearer jwt-token...",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.ResponseBody"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPError"
                        }
                    }
                }
            }
        },
        "/onboarding/step3": {
            "post": {
                "description": "Create or update multiple vehicles for the company in a single request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "onboarding"
                ],
                "summary": "Onboarding Step 3: Create/update vehicles",
                "parameters": [
                    {
                        "description": "Create/update vehicles request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/onboarding.step3Request"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Bearer jwt-token...",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.ResponseBody"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPError"
                        }
                    }
                }
            }
        },
        "/onboarding/step4": {
            "post": {
                "description": "Create or update multiple drivers for the company in a single request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "onboarding"
                ],
                "summary": "Onboarding Step 4: Create/update drivers",
                "parameters": [
                    {
                        "description": "Create/update drivers request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/onboarding.step4Request"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Bearer jwt-token...",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.ResponseBody"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPError"
                        }
                    }
                }
            }
        },
        "/onboarding/step5": {
            "post": {
                "description": "Create initial pricing matrix",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "onboarding"
                ],
                "summary": "Onboarding Step 5: Create pricing matrix",
                "parameters": [
                    {
                        "description": "Create pricing matrix request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/onboarding.step5Request"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Bearer jwt-token...",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.ResponseBody"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "onboarding.driverRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "license_number": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "onboarding.pricingRequest": {
            "type": "object",
            "properties": {
                "customer_id": {
                    "type": "string"
                },
                "dest_city_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "origin_city_id": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                }
            }
        },
        "onboarding.step1Request": {
            "type": "object",
            "required": [
                "company_name",
                "company_type"
            ],
            "properties": {
                "company_name": {
                    "type": "string"
                },
                "company_type": {
                    "type": "string"
                }
            }
        },
        "onboarding.step2Request": {
            "type": "object",
            "properties": {
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/onboarding.userRequest"
                    }
                }
            }
        },
        "onboarding.step3Request": {
            "type": "object",
            "properties": {
                "vehicles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/onboarding.vehicleRequest"
                    }
                }
            }
        },
        "onboarding.step4Request": {
            "type": "object",
            "properties": {
                "drivers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/onboarding.driverRequest"
                    }
                }
            }
        },
        "onboarding.step5Request": {
            "type": "object",
            "properties": {
                "pricing": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/onboarding.pricingRequest"
                    }
                }
            }
        },
        "onboarding.userRequest": {
            "type": "object",
            "properties": {
                "confirm_password": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "onboarding.vehicleRequest": {
            "type": "object",
            "properties": {
                "capacity_volume": {
                    "type": "number"
                },
                "capacity_weight": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "plate_number": {
                    "type": "string"
                },
                "vehicle_type": {
                    "type": "string"
                }
            }
        },
...
        "rest.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "$ref": "#/definitions/rest.Message"
                }
            }
        },
        "rest.Message": {
            "type": "string",
            "enum": [
                "success",
                "resource created",
                "resource updated",
                "resource deleted",
                "invalid request body",
                "missing required fields",
                "invalid field value",
                "validation failed",
                "unauthorized",
                "forbidden",
                "resource not found",
                "conflict",
                "internal server error",
                "service unavailable",
                "invalid request body. please check your input format",
                "method not allowed"
            ],
            "x-enum-varnames": [
                "MsgSuccess",
                "MsgCreated",
                "MsgUpdated",
                "MsgDeleted",
                "MsgInvalidJSON",
                "MsgMissingField",
                "MsgInvalidField",
                "MsgValidationError",
                "MsgUnauthorized",
                "MsgForbidden",
                "MsgNotFound",
                "MsgConflict",
                "MsgInternalError",
                "MsgServiceUnavailable",
                "MsgBadRequest",
                "MsgNotAllowed"
            ]
        },
        "rest.Meta": {
            "type": "object",
            "properties": {
                "has_next": {
                    "type": "boolean"
                },
                "has_prev": {
                    "type": "boolean"
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "rest.ResponseBody": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/rest.Meta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    }
}